+++ scripts/update_offers.py
@@ -18,7 +18,7 @@ non visible.
 Ce script peut uniquement être appelé depuis Flask :
   * Soit depuis l'interface en ligne (menu GRI) ;
   * Soit par ligne de commande :
-    cd /home/intrarez/intrarez; " ./env/bin/flask script update_offers.py
+    cd /home/intrarez/intrarez; ./env/bin/flask script update_offers.py
 
 12/2021 Loïc 137
 """
@@ -35,7 +35,7 @@ except ImportError:
         "  * Soit depuis l'interface en ligne (menu GRI) ;\n"
         "  * Soit par ligne de commande :\n"
         "    cd /home/intrarez/intrarez; "
-        "    ./env/bin/flask script update_sub_states.py\n"
+        "    ./env/bin/flask script update_offers.py\n"
     )
     sys.exit(1)